// src/service/type.ts
export interface PaginationInfo{
    page: number;
    total_pages?: number;
    total_results?: number;
}